# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Talentious Backend API
  version: "1.0.0"
  description: |
    API backend V1 de la plateforme Talentious — génération de CV assistée par IA.

    ## Sécurité
    Tous les endpoints nécessitent un JWT Bearer valide (HS256, 30 min), sauf :
    - `POST /auth/register` et `POST /auth/login` : publics (security: [])
    - `POST /billing/webhook` : public mais authentifié par signature HMAC Stripe
      via l'en-tête `Stripe-Signature` — aucun JWT Talentious n'est attendu.

    ## CareerPass
    Les endpoints de génération de CV vérifient l'existence d'un CareerPass actif
    (table `career_passes`, `valid_until > now()`). Un utilisateur avec le rôle `admin`
    bypasse ce contrôle. En l'absence de pass actif, la réponse est `402` au format
    RFC 7807 (`application/problem+json`).

    ## Erreurs
    Toutes les erreurs suivent le standard RFC 7807 (`application/problem+json`).
    Voir le schéma `Problem` dans `components/schemas`.

    ## Génération asynchrone
    La génération de CV est asynchrone. `POST /cv/generate` retourne `202` avec un
    `job_id`. L'état du job est consultable via `GET /cv/jobs/{job_id}`. Une fois le
    job en état `succeeded`, le CV est accessible via `GET /cv/{cv_id}`. Un canal SSE
    optionnel est disponible sur `GET /cv/jobs/{job_id}/events`.

    Transitions d'état d'un job : queued -> running -> succeeded | failed.
  license:
    name: Proprietary
    url: https://talentious.fr/legal/terms
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Créer un nouveau compte utilisateur
      description: |
        Crée un utilisateur avec le rôle `USER` par défaut et initialise
        un profil vide. Aucun JWT n'est requis. Un email de vérification est
        envoyé ; la connexion est refusée (403) tant que l'adresse n'est pas
        confirmée.

        Anti-squat : si l'email correspond à un compte existant NON vérifié,
        aucune erreur « déjà utilisé » n'est renvoyée — l'email de vérification
        est simplement renvoyé à l'adresse (le compte n'est pas modifié). Le
        titulaire réel de l'adresse récupère le compte via vérification puis
        « mot de passe oublié ». Le 400 « déjà enregistré » n'est renvoyé que
        pour un compte VÉRIFIÉ.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Compte créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Email déjà utilisé
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: "https://talentious.fr/errors/email-already-registered"
                title: "Email Already Registered"
                status: 400
                detail: "An account with this email address already exists."
                instance: "/v1/auth/register"
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: S'authentifier et obtenir un JWT
      description: |
        Authentification par email/mot de passe. Retourne un JWT Bearer valide 30 minutes.
        Le champ `username` du formulaire doit contenir l'adresse email
        (convention OAuth2PasswordRequestForm).
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginForm'
      responses:
        '200':
          description: Authentification réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Identifiants invalides
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: "https://talentious.fr/errors/invalid-credentials"
                title: "Invalid Credentials"
                status: 401
                detail: "Incorrect email or password."
                instance: "/v1/auth/login"
        '403':
          description: |
            Adresse email non vérifiée. Les identifiants sont corrects mais la
            connexion est refusée tant que l'adresse n'a pas été confirmée
            (empêche l'usage d'un compte créé avec l'adresse d'autrui). Le client
            doit proposer le renvoi de l'email (POST /auth/verify-email/resend).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: "https://talentious.fr/errors/email-not-verified"
                title: "Forbidden"
                status: 403
                detail: "Adresse email non vérifiée. Consultez votre boîte mail ou renvoyez l'email de confirmation."
                instance: "/v1/auth/login"
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/verify-email/resend:
    post:
      tags: [Auth]
      operationId: resendVerificationPublic
      summary: Renvoyer l'email de vérification (public)
      description: |
        Renvoie l'email de vérification d'un compte non vérifié. Endpoint PUBLIC
        (écran de connexion, avant authentification). Réponse constante 204
        quelle que soit l'existence du compte (anti-énumération) ; envoi effectif
        uniquement si le compte existe et n'est pas vérifié. Rate-limité.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '204':
          description: Demande acceptée (réponse constante, anti-énumération)
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          description: Trop de demandes (rate limit)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/password/change:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: Changer son mot de passe (connecté)
      description: |
        Change le mot de passe de l'utilisateur authentifié. Exige le mot de
        passe actuel (ré-authentification). Par sécurité, toutes les sessions
        (refresh tokens) sont révoquées et une notification est envoyée par email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '204':
          description: Mot de passe changé, sessions révoquées
        '401':
          description: Mot de passe actuel incorrect (ou non authentifié)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/email/change:
    post:
      tags: [Auth]
      operationId: requestEmailChange
      summary: Demander le changement d'adresse email (connecté)
      description: |
        Envoie un lien de confirmation à la NOUVELLE adresse (l'adresse du compte
        ne change qu'après confirmation — la nouvelle boîte fait autorité). Exige
        le mot de passe actuel. Rate-limité (envoi d'email).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_email, current_password]
              properties:
                new_email:
                  type: string
                  format: email
                current_password:
                  type: string
      responses:
        '204':
          description: Lien de confirmation envoyé à la nouvelle adresse
        '400':
          description: Nouvelle adresse invalide (déjà utilisée ou identique à l'actuelle)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Mot de passe actuel incorrect (ou non authentifié)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Trop de demandes (rate limit)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/email/confirm:
    post:
      tags: [Auth]
      operationId: confirmEmailChange
      summary: Confirmer le changement d'adresse email (lien reçu)
      description: |
        Applique le changement d'adresse depuis le jeton reçu sur la NOUVELLE
        boîte. Endpoint public (le lien est cliqué depuis l'email). L'ancienne
        adresse reçoit une notification de sécurité.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Adresse mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Jeton invalide/expiré, ou adresse devenue indisponible
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Obtenir les informations de l'utilisateur courant
      responses:
        '200':
          description: Informations de l'utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Profile
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT HS256, durée de vie 30 minutes. Obtenu via POST /v1/auth/login.
...
  responses:
    Unauthorized:
      description: JWT absent ou invalide
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://talentious.fr/errors/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Missing or invalid authentication token."
            instance: "/v1/auth/me"
...
    ValidationError:
      description: Erreur de validation du corps de la requête (Pydantic)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://talentious.fr/errors/validation-error"
            title: "Validation Error"
            status: 422
            detail: "Request body validation failed."
            instance: "/v1/profile"

  # ---------------------------------------------------------------------------
  # Schémas
  # ---------------------------------------------------------------------------
  schemas:
...
    Problem:
      type: object
      description: |
        Erreur au format RFC 7807 (application/problem+json).
        Aucune information interne (stack trace, str(e)) ne doit figurer
        dans `detail` en production.
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: URI identifiant le type d'erreur
          example: "https://talentious.fr/errors/career-pass-required"
        title:
          type: string
          description: Titre court lisible par un humain
          example: "Career Pass Required"
        status:
          type: integer
          description: Code HTTP
          example: 402
        detail:
          type: string
          description: Explication détaillée, lisible par un humain (sans internals)
          example: "An active Career Pass is required to generate CVs."
        instance:
          type: string
          description: URI de la ressource ayant provoqué l'erreur
          example: "/v1/cv/generate"

    # =========================================================================
    # Auth
    # =========================================================================
    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: "alice@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: Mot de passe en clair (haché côté serveur avec bcrypt)
          example: "S3cr3t!Pass"
    LoginForm:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          format: email
          description: Adresse email de l'utilisateur (champ `username` par convention OAuth2)
        password:
          type: string
    Token:
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
          description: JWT HS256
        token_type:
          type: string
          enum: [bearer]
          example: "bearer"
    UserResponse:
      type: object
      required:
        - id
        - email
        - role
        - is_active
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
        email_verified:
          type: boolean
          description: L'adresse email a été confirmée (obligatoire pour se connecter)
        display_name:
          type: string
          nullable: true
          description: |
            Nom d'affichage (prénom du profil). Modifiable par l'utilisateur via
            son profil ; null si le profil n'a pas encore de prénom.
        stripe_customer_id:
          type: string
          nullable: true
          description: Identifiant client Stripe (null si jamais passé en caisse)
...